Veronika groups

See merge request dugonogi-progi/radno_vrijeme!17

+++ Backend/backend/src/main/java/progi/dugonogiprogi/radnovrijeme/backend/dao/GroupRepository.java
@@ -12,7 +12,6 @@ import java.util.Optional;
  */
 
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Long> {
-
+public interface GroupRepository extends JpaRepository<Group, Integer> {
 
 }
